create a homepage to the application
